requirements.txt
cloudscraper
beautifulsoup4
requests
brotli
python-dotenv
mysqlclient
